NEST seems to compile under aspnetcore50 now :)
