+++ src/Elasticsearch.Net/Connection/HttpConnection.cs
@@ -12,205 +12,168 @@ using System.Threading.Tasks;
 using Elasticsearch.Net.Connection.Configuration;
 using Elasticsearch.Net.Providers;
 using PurifyNet;
+using System.Net.Http;
+using System.Net.Http.Headers;
 
 namespace Elasticsearch.Net.Connection
 {
-#if !ASPNETCORE50
-	public class HttpConnection2 : IConnection
+	/// <summary>
+	/// IConnection implemented using <see cref="System.Net.Http.HttpClient"/>
+	/// </summary>
+	public class HttpConnection : IConnection, IDisposable
 	{
-		const int BUFFER_SIZE = 1024;
-
-		protected IConnectionConfigurationValues ConnectionSettings { get; set; }
-		private readonly Semaphore _resourceLock;
+		protected readonly IConnectionConfigurationValues _settings;
 
 		public TransportAddressScheme? AddressScheme { get; private set; }
 
-		static HttpConnection2()
+		static HttpConnection()
 		{
-			//ServicePointManager.SetTcpKeepAlive(true, 2000, 2000);
+			// brought over from HttpClient
+			//ServicePointManager.UseNagleAlgorithm = false;
+			//ServicePointManager.Expect100Continue = false;
 
-			//WebException's GetResponse is limitted to 65kb by default.
-			//Elasticsearch can be alot more chatty then that when dumping exceptions
-			//On error responses, so lets up the ante.
-
-			//Not available under mono
-			if (Type.GetType("Mono.Runtime") == null)
-				HttpWebRequest.DefaultMaximumErrorResponseLength = -1;
+			// this should be set globally based on _settings.MaximumAsyncConnections
+			//ServicePointManager.DefaultConnectionLimit = 10000;
 		}
 
-		public HttpConnection2(IConnectionConfigurationValues settings)
+		/// <summary>
+		/// Initializes a new instance of the <see cref="HttpClientConnection"/> class.
+		/// </summary>
+		/// <param name="settings">The settings.</param>
+		/// <param name="handler">The handler.</param>
+		public HttpConnection(IConnectionConfigurationValues settings, HttpClientHandler handler = null)
 		{
-			if (settings == null)
-				throw new ArgumentNullException("settings");
-
+			_settings = settings;
 			if (settings.ConnectionPool.UsingSsl)
 				this.AddressScheme = TransportAddressScheme.Https;
 
-			this.ConnectionSettings = settings;
-			if (settings.MaximumAsyncConnections > 0)
-			{
-				var semaphore = Math.Max(1, settings.MaximumAsyncConnections);
-				this._resourceLock = new Semaphore(semaphore, semaphore);
-			}
-		}
+			DefaultContentType = "application/json";
 
-		public virtual ElasticsearchResponse<Stream> GetSync(Uri uri, IRequestConfiguration requestSpecificConfig = null)
-		{
-			return this.HeaderOnlyRequest(uri, "GET", requestSpecificConfig);
-		}
+			var innerHandler = handler ?? new HttpClientHandler();
 
-		public virtual ElasticsearchResponse<Stream> HeadSync(Uri uri, IRequestConfiguration requestSpecificConfig = null)
+			if (innerHandler.SupportsProxy && !string.IsNullOrWhiteSpace(_settings.ProxyAddress))
 		{
-			return this.HeaderOnlyRequest(uri, "HEAD", requestSpecificConfig);
-		}
-
-		public virtual ElasticsearchResponse<Stream> PostSync(Uri uri, byte[] data, IRequestConfiguration requestSpecificConfig = null)
-		{
-			return this.BodyRequest(uri, data, "POST", requestSpecificConfig);
-		}
-		
-		public virtual ElasticsearchResponse<Stream> PutSync(Uri uri, byte[] data, IRequestConfiguration requestSpecificConfig = null)
-		{
-			return this.BodyRequest(uri, data, "PUT", requestSpecificConfig);
-		}
-		
-		public virtual ElasticsearchResponse<Stream> DeleteSync(Uri uri, IRequestConfiguration requestSpecificConfig = null)
-		{
-			return this.HeaderOnlyRequest(uri, "DELETE", requestSpecificConfig);
-		}
-		
-		public virtual ElasticsearchResponse<Stream> DeleteSync(Uri uri, byte[] data, IRequestConfiguration requestSpecificConfig = null)
-		{
-			return this.BodyRequest(uri, data, "DELETE", requestSpecificConfig);
-		}
-
-		private ElasticsearchResponse<Stream> HeaderOnlyRequest(Uri uri, string method, IRequestConfiguration requestSpecificConfig)
-		{
-			var r = this.CreateHttpWebRequest(uri, method, null, requestSpecificConfig);
-			return this.DoSynchronousRequest(r, requestSpecificConfig: requestSpecificConfig);
-		}
-
-		private ElasticsearchResponse<Stream> BodyRequest(Uri uri, byte[] data, string method, IRequestConfiguration requestSpecificConfig)
-		{
-			var r = this.CreateHttpWebRequest(uri, method, data, requestSpecificConfig);
-			return this.DoSynchronousRequest(r, data, requestSpecificConfig);
-		}
-
-		public virtual Task<ElasticsearchResponse<Stream>> Get(Uri uri, IRequestConfiguration requestSpecificConfig = null)
-		{
-			var r = this.CreateHttpWebRequest(uri, "GET", null, requestSpecificConfig);
-			return this.DoAsyncRequest(r, requestSpecificConfig: requestSpecificConfig);
-		}
-		
-		public virtual Task<ElasticsearchResponse<Stream>> Head(Uri uri, IRequestConfiguration requestSpecificConfig = null)
-		{
-			var r = this.CreateHttpWebRequest(uri, "HEAD", null, requestSpecificConfig);
-			return this.DoAsyncRequest(r, requestSpecificConfig: requestSpecificConfig);
-		}
-		
-		public virtual Task<ElasticsearchResponse<Stream>> Post(Uri uri, byte[] data, IRequestConfiguration requestSpecificConfig = null)
-		{
-			var r = this.CreateHttpWebRequest(uri, "POST", data, requestSpecificConfig);
-			return this.DoAsyncRequest(r, data, requestSpecificConfig: requestSpecificConfig);
-		}
-
-		public virtual Task<ElasticsearchResponse<Stream>> Put(Uri uri, byte[] data, IRequestConfiguration requestSpecificConfig = null)
-		{
-			var r = this.CreateHttpWebRequest(uri, "PUT", data, requestSpecificConfig);
-			return this.DoAsyncRequest(r, data, requestSpecificConfig: requestSpecificConfig);
-		}
-
-		public virtual Task<ElasticsearchResponse<Stream>> Delete(Uri uri, byte[] data, IRequestConfiguration requestSpecificConfig = null)
-		{
-			var r = this.CreateHttpWebRequest(uri, "DELETE", data, requestSpecificConfig);
-			return this.DoAsyncRequest(r, data, requestSpecificConfig: requestSpecificConfig);
-		}
-		
-		public virtual Task<ElasticsearchResponse<Stream>> Delete(Uri uri, IRequestConfiguration requestSpecificConfig = null)
-		{
-			var r = this.CreateHttpWebRequest(uri, "DELETE", null, requestSpecificConfig);
-			return this.DoAsyncRequest(r, requestSpecificConfig: requestSpecificConfig);
-		}
-
-		private static void ThreadTimeoutCallback(object state, bool timedOut)
-		{
-			if (timedOut)
-			{
-				HttpWebRequest request = state as HttpWebRequest;
-				if (request != null)
+				innerHandler.Proxy = new ClientConnectionProxy(new Uri(_settings.ProxyAddress))
 				{
-					request.Abort();
+					Credentials = new NetworkCredential(_settings.ProxyUsername, _settings.ProxyPassword)
+				};
+
+				innerHandler.UseProxy = true;
+		}
+
+			Client = new HttpClient(new InternalHttpMessageHandler(innerHandler), false)
+		{
+				Timeout = TimeSpan.FromMilliseconds(_settings.Timeout)
+			};
+			if (settings.EnableCompressedResponses ||
+			    settings.EnableHttpCompression && innerHandler.SupportsAutomaticDecompression)
+		{
+				innerHandler.AutomaticDecompression = DecompressionMethods.GZip | DecompressionMethods.Deflate;
+				Client.DefaultRequestHeaders.AcceptEncoding.Add(new StringWithQualityHeaderValue("gzip"));
+				Client.DefaultRequestHeaders.AcceptEncoding.Add(new StringWithQualityHeaderValue("deflate"));
+		}
+
+		}
+
+		/// <summary>
+		/// Gets or sets the default type of the content.
+		/// </summary>
+		/// <value>The default type of the content.</value>
+		public string DefaultContentType { get; set; }
+
+		/// <summary>
+		/// Gets a value indicating whether this instance is disposed.
+		/// </summary>
+		/// <value><c>true</c> if this instance is disposed; otherwise, <c>false</c>.</value>
+		public bool IsDisposed { get; private set; }
+
+		/// <summary>
+		/// Gets the client.
+		/// </summary>
+		/// <value>The client.</value>
+		public HttpClient Client { get; private set; }
+
+		/// <summary>
+		/// Wraps the DoRequest to run synchronously
+		/// </summary>
+		/// <param name="method">The method.</param>
+		/// <param name="uri">The URI.</param>
+		/// <param name="data">The data.</param>
+		/// <param name="requestSpecificConfig">The request specific configuration.</param>
+		/// <returns>ElasticsearchResponse&lt;Stream&gt;.</returns>
+		public virtual ElasticsearchResponse<Stream> DoRequestSync(HttpMethod method, Uri uri, byte[] data = null, IRequestConfiguration requestSpecificConfig = null)
+		{
+			ThrowIfDisposed();
+
+			try
+		{
+				return this.DoRequestInternal(method, uri, data, requestSpecificConfig)
+					.Result;
+		}
+			catch (AggregateException ex)
+		{
+				return ElasticsearchResponse<Stream>.CreateError(_settings, ex.Flatten(), method.ToString().ToLowerInvariant(),
+					uri.ToString(), data);
+		}
+			catch (Exception ex)
+		{
+				return ElasticsearchResponse<Stream>.CreateError(_settings, ex, method.ToString().ToLowerInvariant(), uri.ToString(),
+					data);
 				}
 			}
+
+		/// <summary>
+		/// Makes an async call to the specified url. Uses the timeout from the IRequestSpecifiConfig is supplied, or the global timeout from settings.
+		/// </summary>
+		/// <param name="method">The method.</param>
+		/// <param name="uri">The URI.</param>
+		/// <param name="data">The data.</param>
+		/// <param name="requestSpecificConfig">The request specific configuration.</param>
+		/// <returns>Task&lt;ElasticsearchResponse&lt;Stream&gt;&gt;.</returns>
+		public virtual async Task<ElasticsearchResponse<Stream>> DoRequest(HttpMethod method, Uri uri, byte[] data = null,
+			IRequestConfiguration requestSpecificConfig = null)
+		{
+			return await this.DoRequestInternal(method, uri, data, requestSpecificConfig).ConfigureAwait(false);
 		}
 
-		protected virtual void AlterServicePoint(ServicePoint requestServicePoint)
+		public async Task<ElasticsearchResponse<Stream>> DoRequestInternal(
+			HttpMethod method, Uri uri, byte[] data = null, IRequestConfiguration requestSpecificConfig = null)
 		{
-			requestServicePoint.UseNagleAlgorithm = false;
-			requestServicePoint.Expect100Continue = false;
-			requestServicePoint.ConnectionLimit = 10000;
-			//looking at http://referencesource.microsoft.com/#System/net/System/Net/ServicePoint.cs
-			//this method only sets internal values and wont actually cause timers and such to be reset
-			//So it should be idempotent if called with the same parameters
-			if (this.ConnectionSettings.KeepAliveTime.HasValue && this.ConnectionSettings.KeepAliveInterval.HasValue)
-				requestServicePoint.SetTcpKeepAlive(true, this.ConnectionSettings.KeepAliveTime.Value, this.ConnectionSettings.KeepAliveInterval.Value);
-		}
+			ThrowIfDisposed();
 
-		protected virtual HttpWebRequest CreateHttpWebRequest(Uri uri, string method, byte[] data, IRequestConfiguration requestSpecificConfig)
+			try
 		{
-			var request = this.CreateWebRequest(uri, method, data, requestSpecificConfig);
-			this.SetBasicAuthenticationIfNeeded(uri, request, requestSpecificConfig);
-			this.SetProxyIfNeeded(request);
-			this.AlterServicePoint(request.ServicePoint);
-			return request;
-		}
+				var request = new HttpRequestMessage(method, uri);
+				if (requestSpecificConfig != null && !string.IsNullOrWhiteSpace(requestSpecificConfig.ContentType))
+					request.Headers.Accept.Add(new MediaTypeWithQualityHeaderValue(requestSpecificConfig.ContentType));
+				else if (!string.IsNullOrWhiteSpace(DefaultContentType))
+					request.Headers.Accept.Add(new MediaTypeWithQualityHeaderValue(DefaultContentType));
 
-		private void SetProxyIfNeeded(HttpWebRequest myReq)
-		{
-			if (!string.IsNullOrEmpty(this.ConnectionSettings.ProxyAddress))
+				if (!string.IsNullOrEmpty(uri.UserInfo))
 			{
-				var proxy = new WebProxy();
-				var uri = new Uri(this.ConnectionSettings.ProxyAddress);
-				var credentials = new NetworkCredential(this.ConnectionSettings.ProxyUsername, this.ConnectionSettings.ProxyPassword);
-				proxy.Address = uri;
-				proxy.Credentials = credentials;
-				myReq.Proxy = proxy;
+					request.Headers.Authorization = new AuthenticationHeaderValue("Basic",
+						Convert.ToBase64String(Encoding.UTF8.GetBytes(uri.UserInfo)));
 			}
 
-			if (this.ConnectionSettings.DisableAutomaticProxyDetection)
+				if (method != HttpMethod.Get && method != HttpMethod.Head && data != null && data.Length > 0)
 			{
-				myReq.Proxy = null;
+					request.Content = new ByteArrayContent(data);
+					if (!string.IsNullOrWhiteSpace(DefaultContentType) && request.Content != null && request.Content.Headers != null)
+						request.Content.Headers.ContentType = new MediaTypeHeaderValue(DefaultContentType);
 			}
+
+				var response = await Client.SendAsync(request, HttpCompletionOption.ResponseHeadersRead)
+					.ConfigureAwait(false);
+
+				if (method == HttpMethod.Head || response.Content == null)
+		{
+					return ElasticsearchResponse<Stream>.Create(_settings, (int) response.StatusCode,
+						method.ToString().ToLowerInvariant(), uri.ToString(), data);
 		}
 
-		private void SetBasicAuthenticationIfNeeded(Uri uri, HttpWebRequest request, IRequestConfiguration requestSpecificConfig)
-		{
-			// Basic auth credentials take the following precedence (highest -> lowest):
-			// 1 - Specified on the request (highest precedence)
-			// 2 - Specified at the global IConnectionSettings level
-			// 3 - Specified with the URI (lowest precedence)
-
-			var userInfo = Uri.UnescapeDataString(uri.UserInfo);
-
-			if (this.ConnectionSettings.BasicAuthorizationCredentials != null)
-				userInfo = this.ConnectionSettings.BasicAuthorizationCredentials.ToString();
-
-			if (requestSpecificConfig != null && requestSpecificConfig.BasicAuthorizationCredentials != null)
-				userInfo = requestSpecificConfig.BasicAuthorizationCredentials.ToString();
-
-			if (!userInfo.IsNullOrEmpty())
-				request.Headers["Authorization"] = "Basic " + Convert.ToBase64String(Encoding.UTF8.GetBytes(userInfo));
-		}
-
-		protected virtual HttpWebRequest CreateWebRequest(Uri uri, string method, byte[] data, IRequestConfiguration requestSpecificConfig)
-		{
-			var request = (HttpWebRequest)WebRequest.Create(uri);
-			request.Accept = "application/json";
-			request.ContentType = "application/json";
-			request.MaximumResponseHeadersLength = -1;
-			request.ServicePoint.SetTcpKeepAlive(true, 2000, 2000);
-			request.Pipelined = this.ConnectionSettings.HttpPipeliningEnabled
-				|| (requestSpecificConfig != null && requestSpecificConfig.EnableHttpPipelining);
+				var responseStream = await response.Content.ReadAsStreamAsync();
+				return ElasticsearchResponse<Stream>.Create(_settings, (int) response.StatusCode,
+					method.ToString().ToLowerInvariant(), uri.ToString(), data, responseStream);
 
 			if (this.ConnectionSettings.EnableHttpCompression)
 			{
@@ -219,214 +182,108 @@ namespace Elasticsearch.Net.Connection
 				if (this.ConnectionSettings.EnableHttpCompression)
 					request.Headers.Add("Content-Encoding", "gzip");
 			}
-
-			if (requestSpecificConfig != null && !string.IsNullOrWhiteSpace(requestSpecificConfig.ContentType))
+			catch (Exception ex)
 			{
-				request.Accept = requestSpecificConfig.ContentType;
-				request.ContentType = requestSpecificConfig.ContentType;
+				return ElasticsearchResponse<Stream>.CreateError(_settings, ex, method.ToString().ToLowerInvariant(), uri.ToString(),
+					data);
 			}
-
-			var timeout = GetRequestTimeout(requestSpecificConfig);
-			request.Timeout = timeout;
-			request.ReadWriteTimeout = timeout;
-			request.Method = method;
-
-			//WebRequest won't send Content-Length: 0 for empty bodies
-			//which goes against RFC's and might break i.e IIS when used as a proxy.
-			//see: https://github.com/elasticsearch/elasticsearch-net/issues/562
-			var m = method.ToLowerInvariant();
-			if (m != "head" && m != "get" && (data == null || data.Length == 0))
-				request.ContentLength = 0;
-
-			return request;
 		}
 
-		protected virtual ElasticsearchResponse<Stream> DoSynchronousRequest(HttpWebRequest request, byte[] data = null, IRequestConfiguration requestSpecificConfig = null)
+		Task<ElasticsearchResponse<Stream>> IConnection.Get(Uri uri, IRequestConfiguration requestSpecificConfig)
 		{
-			var path = request.RequestUri.ToString();
-			var method = request.Method;
+			return DoRequest(HttpMethod.Get, uri, null, requestSpecificConfig);
+		}
 
-			if (data != null)
+		ElasticsearchResponse<Stream> IConnection.GetSync(Uri uri, IRequestConfiguration requestSpecificConfig)
 			{
+			return DoRequestSync(HttpMethod.Get, uri, null, requestSpecificConfig);
+		}
 
-				using (var r = request.GetRequestStream())
+		Task<ElasticsearchResponse<Stream>> IConnection.Head(Uri uri, IRequestConfiguration requestSpecificConfig)
 				{
-					if (this.ConnectionSettings.EnableHttpCompression)
-						using (var zipStream = new GZipStream(r, CompressionMode.Compress))
-							zipStream.Write(data, 0, data.Length);
-					else 
-						r.Write(data, 0, data.Length);
+			return DoRequest(HttpMethod.Head, uri, null, requestSpecificConfig);
 				}
-			}
-			try
+
+		ElasticsearchResponse<Stream> IConnection.HeadSync(Uri uri, IRequestConfiguration requestSpecificConfig)
 			{
-				//http://msdn.microsoft.com/en-us/library/system.net.httpwebresponse.getresponsestream.aspx
-				//Either the stream or the response object needs to be closed but not both although it won't
-				//throw any errors if both are closed atleast one of them has to be Closed.
-				//Since we expose the stream we let closing the stream determining when to close the connection
-				var response = (HttpWebResponse)request.GetResponse();
-				var responseStream = response.GetResponseStream();
-				return WebToElasticsearchResponse(data, responseStream, response, method, path);
+			return DoRequestSync(HttpMethod.Head, uri, null, requestSpecificConfig);
 			}
-			catch (WebException webException)
+
+		Task<ElasticsearchResponse<Stream>> IConnection.Post(Uri uri, byte[] data, IRequestConfiguration requestSpecificConfig)
 			{
-				return HandleWebException(data, webException, method, path);
+			return DoRequest(HttpMethod.Post, uri, data, requestSpecificConfig);
 			}
+
+		ElasticsearchResponse<Stream> IConnection.PostSync(Uri uri, byte[] data, IRequestConfiguration requestSpecificConfig)
+		{
+			return DoRequestSync(HttpMethod.Post, uri, data, requestSpecificConfig);
+			}
+
+		Task<ElasticsearchResponse<Stream>> IConnection.Put(Uri uri, byte[] data, IRequestConfiguration requestSpecificConfig)
+		{
+			return DoRequest(HttpMethod.Put, uri, data, requestSpecificConfig);
 		}
 
-		private ElasticsearchResponse<Stream> HandleWebException(byte[] data, WebException webException, string method, string path)
+		ElasticsearchResponse<Stream> IConnection.PutSync(Uri uri, byte[] data, IRequestConfiguration requestSpecificConfig)
 		{
-			ElasticsearchResponse<Stream> cs = null;
-			var httpEx = webException.Response as HttpWebResponse;
-			if (httpEx != null)
-			{
-				//StreamReader ms = new StreamReader(httpEx.GetResponseStream());
-				//var response = ms.ReadToEnd();
-				cs = WebToElasticsearchResponse(data, httpEx.GetResponseStream(), httpEx, method, path);
-				return cs;
-			}
-			cs = ElasticsearchResponse<Stream>.CreateError(this.ConnectionSettings, webException, method, path, data);
-			return cs;
+			return DoRequestSync(HttpMethod.Put, uri, data, requestSpecificConfig);
 		}
 
-		private ElasticsearchResponse<Stream> WebToElasticsearchResponse(byte[] data, Stream responseStream, HttpWebResponse response, string method, string path)
+		Task<ElasticsearchResponse<Stream>> IConnection.Delete(Uri uri, IRequestConfiguration requestSpecificConfig)
+			{
+			return DoRequest(HttpMethod.Delete, uri, null, requestSpecificConfig);
+			}
+
+		ElasticsearchResponse<Stream> IConnection.DeleteSync(Uri uri, IRequestConfiguration requestSpecificConfig)
+			{
+			return DoRequestSync(HttpMethod.Delete, uri, null, requestSpecificConfig);
+			}
+
+		Task<ElasticsearchResponse<Stream>> IConnection.Delete(Uri uri, byte[] data,
+			IRequestConfiguration requestSpecificConfig)
+			{
+			return DoRequest(HttpMethod.Delete, uri, data, requestSpecificConfig);
+			}
+
+		ElasticsearchResponse<Stream> IConnection.DeleteSync(Uri uri, byte[] data, IRequestConfiguration requestSpecificConfig)
 		{
-			ElasticsearchResponse<Stream> cs = ElasticsearchResponse<Stream>.Create(this.ConnectionSettings, (int)response.StatusCode, method, path, data);
-			cs.Response = responseStream;
-			return cs;
+			return DoRequestSync(HttpMethod.Delete, uri, data, requestSpecificConfig);
 		}
 
-		protected virtual Task<ElasticsearchResponse<Stream>> DoAsyncRequest(HttpWebRequest request, byte[] data = null, IRequestConfiguration requestSpecificConfig = null)
+		private void ThrowIfDisposed()
 		{
-			var tcs = new TaskCompletionSource<ElasticsearchResponse<Stream>>();
-			if (this.ConnectionSettings.MaximumAsyncConnections <= 0
-			  || this._resourceLock == null)
-				return this.CreateIterateTask(request, data, requestSpecificConfig, tcs);
-
-			var timeout = GetRequestTimeout(requestSpecificConfig);
-			var path = request.RequestUri.ToString();
-			var method = request.Method;
-			if (!this._resourceLock.WaitOne(timeout))
+			if (IsDisposed)
 			{
-				var m = "Could not start the operation before the timeout of " + timeout +
-				  "ms completed while waiting for the semaphore";
-				var cs = ElasticsearchResponse<Stream>.CreateError(this.ConnectionSettings, new TimeoutException(m), method, path, data);
-				tcs.SetResult(cs);
-				return tcs.Task;
-			}
-			try
-			{
-				return this.CreateIterateTask(request, data, requestSpecificConfig, tcs);
-			}
-			finally
-			{
-				this._resourceLock.Release();
-			}
-		}
-
-		private Task<ElasticsearchResponse<Stream>> CreateIterateTask(HttpWebRequest request, byte[] data, IRequestConfiguration requestSpecificConfig, TaskCompletionSource<ElasticsearchResponse<Stream>> tcs)
-		{
-			this.Iterate(request, data, this._AsyncSteps(request, tcs, data, requestSpecificConfig), tcs);
-			return tcs.Task;
-		}
-
-		private IEnumerable<Task> _AsyncSteps(HttpWebRequest request, TaskCompletionSource<ElasticsearchResponse<Stream>> tcs, byte[] data, IRequestConfiguration requestSpecificConfig)
-		{
-			var timeout = GetRequestTimeout(requestSpecificConfig);
-
-			if (data != null)
-			{
-				var getRequestStream = Task.Factory.FromAsync<Stream>(request.BeginGetRequestStream, request.EndGetRequestStream, null);
-				ThreadPool.RegisterWaitForSingleObject((getRequestStream as IAsyncResult).AsyncWaitHandle, ThreadTimeoutCallback, request, timeout, true);
-				yield return getRequestStream;
-
-				var requestStream = getRequestStream.Result;
-				try
-				{
-					if (this.ConnectionSettings.EnableHttpCompression)
-					{
-						using (var zipStream = new GZipStream(requestStream, CompressionMode.Compress))
-						{
-
-							var writeToRequestStream = Task.Factory.FromAsync(zipStream.BeginWrite, zipStream.EndWrite, data, 0,
-								data.Length, null);
-							yield return writeToRequestStream;
+				throw new ObjectDisposedException(GetType().Name);
 						}
 					}
-					else
+
+		public void Dispose()
 					{
-						var writeToRequestStream = Task.Factory.FromAsync(requestStream.BeginWrite, requestStream.EndWrite, data, 0,
-							data.Length, null);
-						yield return writeToRequestStream;
+			Dispose(true);
+			GC.SuppressFinalize(this);
 					}
-				}
-				finally
+
+		~HttpConnection()
 				{
-					requestStream.Close();
+			Dispose(false);
 				}
-			}
 
-			// Get the response
-			var getResponse = Task.Factory.FromAsync<WebResponse>(request.BeginGetResponse, request.EndGetResponse, null);
-			ThreadPool.RegisterWaitForSingleObject((getResponse as IAsyncResult).AsyncWaitHandle, ThreadTimeoutCallback, request, timeout, true);
-			yield return getResponse;
+		protected virtual void Dispose(bool disposing)
+		{
+			if (IsDisposed)
+				return;
 
-			var path = request.RequestUri.ToString();
-			var method = request.Method;
-
-			//http://msdn.microsoft.com/en-us/library/system.net.httpwebresponse.getresponsestream.aspx
-			//Either the stream or the response object needs to be closed but not both (although it won't)
-			//throw any errors if both are closed atleast one of them has to be Closed.
-			//Since we expose the stream we let closing the stream determining when to close the connection
-			var response = (HttpWebResponse)getResponse.Result;
-			var responseStream = response.GetResponseStream();
-			var cs = ElasticsearchResponse<Stream>.Create(this.ConnectionSettings, (int)response.StatusCode, method, path, data);
-			cs.Response = responseStream;
-			tcs.TrySetResult(cs);
+			if (disposing)
+					{
+				if (Client != null)
+				{
+					Client.Dispose();
+					Client = null;
+				}
 		}
 
-		private void Iterate(HttpWebRequest request, byte[] data, IEnumerable<Task> asyncIterator, TaskCompletionSource<ElasticsearchResponse<Stream>> tcs)
-		{
-			var enumerator = asyncIterator.GetEnumerator();
-			Action<Task> recursiveBody = null;
-			recursiveBody = completedTask =>
-			{
-				if (completedTask != null && completedTask.IsFaulted)
-				{
-					//none of the individual steps in _AsyncSteps run in parallel for 1 request
-					//as this would be impossible we can assume Aggregate Exception.InnerException
-					var exception = completedTask.Exception.InnerException;
-
-					//cleanly exit from exceptions in stages if the exception is a webexception
-					if (exception is WebException)
-					{
-						var path = request.RequestUri.ToString();
-						var method = request.Method;
-						var response = this.HandleWebException(data, exception as WebException, method, path);
-						tcs.SetResult(response);
-					}
-					else
-						tcs.TrySetException(exception);
-					enumerator.Dispose();
-				}
-				else if (enumerator.MoveNext())
-				{
-					enumerator.Current.ContinueWith(recursiveBody, TaskContinuationOptions.ExecuteSynchronously);
-				}
-				else enumerator.Dispose();
-			};
-			recursiveBody(null);
-		}
-
-		private int GetRequestTimeout(IRequestConfiguration requestConfiguration)
-		{
-			if (requestConfiguration != null && requestConfiguration.RequestTimeout.HasValue)
-				return requestConfiguration.RequestTimeout.Value;
-
-			return this.ConnectionSettings.Timeout;
+			IsDisposed = true;
 		}
 	}
-
-#endif
-
 }
