+++ src/Elasticsearch.Net/Domain/Enums.Generated.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Runtime.Serialization;
 
+
 ///This file contains all the typed enums that the client rest api spec exposes.
 ///This file is automatically generated from https://github.com/elasticsearch/elasticsearch-rest-api-spec
 ///Generated of commit
