+++ src/Nest/DSL/Query/BoostingQueryDescriptor.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Configuration;
 using System.Linq;
 using Nest.Resolvers.Converters;
 using Newtonsoft.Json;
