+++ src/Nest/Domain/Responses/GetFieldMappingResponse.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Configuration;
 using System.Linq;
 using System.Linq.Expressions;
 using Elasticsearch.Net;
