+++ src/Nest/Extensions/StringExtensions.cs
@@ -34,7 +34,12 @@ namespace Nest
 			if (!char.IsUpper(s[0]))
 				return s;
 
+#if ASPNETCORE50
+			//TODO write proper unit test for this
+			string camelCase = char.ToLowerInvariant(s[0]).ToString();
+#else
 			string camelCase = char.ToLower(s[0], CultureInfo.InvariantCulture).ToString(CultureInfo.InvariantCulture);
+#endif
 			if (s.Length > 1)
 				camelCase += s.Substring(1);
